% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{add_factor_order}
\alias{add_factor_order}
\title{Add \code{hclust} order to columns.}
\usage{
add_factor_order(tbl, col1, col2, value, sym = FALSE)
}
\arguments{
\item{tbl}{A tibble.}

\item{col1, col2}{Columns to add order to.}

\item{value}{Column giving a value from each pair of levels in \code{col1}
and \code{col2}.}

\item{sym}{Is the relationship between \code{col1} and \code{col2} symetric.
Meaning they represent the same physical property and the expanded matrix of
value pairs will be symetrical across the diagonal.}
}
\description{
Add orders to two character columns in a tibble, based on heirarchical
clustering of a third column.
}
