% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{alphabetise_matrix}
\alias{alphabetise_matrix}
\title{Sort matric columns and/or rows alphabetically}
\usage{
alphabetise_matrix(x, by = c("rows", "columns", "both"))
}
\arguments{
\item{x}{A matrix.}

\item{by}{Whether to sort by rows, columns or both. Supports argument
matching on any unique substring.}
}
\description{
Sort matric columns and/or rows alphabetically
}
