% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{col_unique_counts}
\alias{col_unique_counts}
\title{Count the number of unique entries in each column}
\usage{
col_unique_counts(tbl)
}
\arguments{
\item{tbl}{A data frame.}
}
\description{
Shortcut for applying \code{\link[dplyr]{n_distinct}} to each tibble column.
}
