% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{enumerate_unique_rows}
\alias{enumerate_unique_rows}
\title{Enumerate unique and duplicate rows in a matrix like object.}
\usage{
enumerate_unique_rows(mat)
}
\arguments{
\item{mat}{A matrix like object}
}
\value{
A list containing the a vector of unique row indices
    \code{indeces}, corresponding to each rows coresponding unique row in
    a deduplicated matrix, and a logical vector showing whether each row is
    duplicate (\code{duplicated}). These two vectors are sufficient to
    extract a  deduplcated matrix (\code{deduped <- mat[!duplicate,]}) and
    reconstruct  the original (\code{mat == deduped[indeces,]}).
}
\description{
Identify which rows of a matrix are unique and which are duplicates,
including identifying which row they are a duplicate of. Unique rows are
numbered sequentially.
}
