% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_tibbles.R
\name{tibble_pca}
\alias{tibble_pca}
\title{Calculate a PCA from selected columns in a tibble}
\usage{
tibble_pca(tbl, ...)
}
\arguments{
\item{tbl}{A tibble.}

\item{...}{Tidy selection of columns to pass to \code{prcomp}}
}
\description{
Use \code{prcomp} to generate a principal component analysis from a
selection of columns in a tibble.
}
