% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_tibbles.R
\name{tibble_to_matrix}
\alias{tibble_to_matrix}
\title{Convert columns from a tibble into a matrix.}
\usage{
tibble_to_matrix(tbl, ..., row_names = NULL)
}
\arguments{
\item{tbl}{Tibble to covert.}

\item{...}{Tidy selection of columns to include in matrix.}

\item{row_names}{String of column name to pull row names from or a character
vector of row names.}
}
\description{
Convert subsections of tibbles into matrices, with the option of adding
row names from another column or a character vector
}
