% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_tibbles.R
\name{tibble_tsne}
\alias{tibble_tsne}
\title{Calculate a tSNE from selected columns}
\usage{
tibble_tsne(tbl, ..., tsne_kwargs = list())
}
\arguments{
\item{tbl}{A tibble.}

\item{...}{A tidy selection of columns}

\item{tsne_kwargs}{Additional arguments to pass to \code{Rtsne}}
}
\description{
Pass a selection of columns to \code{\link[Rtsne]{Rtnse}}, deduplicating
them and then reconstructing the projection for duplicate rows.
}
