% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_tibbles.R
\name{transpose_tibble}
\alias{transpose_tibble}
\title{Transpose a tibble.}
\usage{
transpose_tibble(tbl, col_names, id_col = "columns")
}
\arguments{
\item{tbl}{Tibble to transpose.}

\item{col_names}{Column containing the transposed column headers.}

\item{id_col}{Name of the new column containing the old column headers.}
}
\description{
Transpose tibble, adding a new column to store the old column names and
drawing new column names from a given column. Currently doesn't have type
checking so new types have to be assigned to columns manually if there
is more than one type in the tibble.
}
