% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_tibbles.R
\name{tibble_correlation}
\alias{tibble_correlation}
\title{Calculate correlations between tibble columns}
\usage{
tibble_correlation(tbl, x, y = NULL, filter_diag = FALSE, ...)
}
\arguments{
\item{tbl}{A tibble.}

\item{x}{Tidy selection of variables for x}

\item{y}{Optional tidy selection of variables for y. If given correlations
between x and y are calculated, otherwise all correlations between x are
calculated.}

\item{filter_diag}{Whether to filter diagonal entries}

\item{...}{Other arguments to pass to \code{cor}}
}
\description{
Calculate the correlation between all pairs from a selection of columns.
The column names are stored in factor columns \code{cat1} and \code{cat2},
with order based upon \code{hclust}
}
